requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
torch==2.0.1
transformers==4.33.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
pandas==2.1.0
numpy==1.25.2
tqdm==4.66.1
